requirement.txt
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Django==4.1.7
djangorestframework==3.14.0
et-xmlfile==1.1.0
greenlet==2.0.2
gunicorn==20.1.0
mysqlclient==2.1.1
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.1
six==1.16.0
SQLAlchemy==2.0.6
sqlparse==0.4.3
typing-extensions==4.5.0
tzdata==2022.7
